careerpath_project/requirements.txt
fastapi
passlib[bcrypt]
uvicorn
gunicorn
python-dotenv
SQLAlchemy
psycopg2-binary
sib-api-v3-sdk
google-api-python-client
google-auth-oauthlib
PyPDF2
python-docx
email-validator
google.generativeai
python-jose
pandas
python-multipart
cloudinary
